KnowledgeTracing/Data-Extraction/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
async-generator==1.10
attrs==17.4.0
Automat==0.6.0
backcall==0.2.0
bleach==4.1.0
blinker==1.4
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
click==6.7
cloud-init==22.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==4.1.2
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.4
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipykernel==5.5.6
ipython==7.16.3
ipython_genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
MarkupSafe==1.0
matplotlib==3.3.4
mistune==0.8.4
mysql-connector-python==8.0.31
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.4.10
numpy==1.19.5
oauthlib==2.0.6
packaging==21.3
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
Pygments==2.13.0
PyGObject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==3.0.9
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2022.6
pyxdg==0.25
PyYAML==3.12
pyzmq==24.0.1
qtconsole==5.2.2
QtPy==2.0.1
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.11.0
sos==4.3
ssh-import-id==5.7
systemd-python==234
terminado==0.12.1
testpath==0.6.0
tornado==6.1
traitlets==4.3.2
Twisted==17.9.0
typing_extensions==4.1.1
ubuntu-advantage-tools==27.10
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
zope.interface==4.3.2
